// src/models/product_header.ts
export interface ProductHeader {
    id?: number;
    lang_code?: string;
    page_code?: string;
    head_text?: string;
    head_textclr?: string;
    shrt_desc?: string;
    shrt_clor?: string;
    tagx_text?: string;
    tagx_clor?: string;
    icon_prod?: string;
    imge_webx?: string;
    imge_tabl?: string;
    imge_mobi?: string;
    bg_clor?: string;
    type_header?: number;
    createdAt?: Date;
    updatedAt?: null;
}